updated jcompilo

+++ jcompilo.sh
@@ -1,22 +1,23 @@
 #!/bin/sh
 
-version=93
-name=jcompilo
+version=99
+artifact=jcompilo
+group=com/googlecode/${artifact}
+repo=repo.bodar.com.s3.amazonaws.com
 dir=lib/
-jar=${dir}${name}.jar
-url=http://${name}.googlecode.com/files/${name}-${version}.jar
+jar=${dir}${artifact}.jar
+url=http://${repo}/${group}/${artifact}/${version}/${artifact}-${version}
+remote_jar=${url}.jar
+remote_sh=${url}.sh
 
-while getopts "u" opt; do
-  case $opt in
-    u) 
-      rm ${jar}
-      shift $(optind)
-      ;;
-  esac
-done
+if [ $1 = "-u" ]; then 
+	rm ${jar}
+	shift 1
+fi
 
 if [ ! -f ${jar} ]; then
 	mkdir -p ${dir} 
-	wget ${url} -O ${jar}
+	wget -O ${jar} ${remote_jar} || curl -o ${jar} ${remote_jar}
+	wget -O $0 ${remote_sh} || curl -o $0 ${remote_sh}
 fi
 exec java -jar ${jar} $*
